chore: improve refactor effort
